+++ src/server/app.js
@@ -1,0 +1,22 @@
+require('env2')('config.env');
+const path = require('path');
+const express = require('express');
+const app = express();
+const routes = require('./controllers/routes');
+
+require('./middlewares/appMiddleware')(app, express);
+
+app.use((err, req, res, next) => {
+  console.error(err.stack);
+  res.status(500).send('Something broke!');
+  next();
+});
+
+app.use(express.static('public'));
+app.use('/api/v1/', routes);
+app.use('/api/v1/', (req, res) => {
+  res.status(404).json({ message: 'page  not found' });
+});
+app.use('/*', express.static(path.join(__dirname,'..','..','public','index.html')));
+
+module.exports = app;
