+++ src/server/index.js
@@ -1,23 +1,7 @@
-require('env2')('config.env');
-const path = require('path');
-const express = require('express');
-const app = express();
-const routes = require('./controllers/routes');
+const http = require('http');
+const app = require('./app.js');
 
-require('./middlewares/appMiddleware')(app, express);
-
-app.use((err, req, res, next) => {
-  console.error(err.stack);
-  res.status(500).send('Something broke!');
-  next();
-});
-
-app.use(express.static('public'));
-app.use('/api/v1/', routes);
-app.use('/api/v1/', (req, res) => {
-  res.status(404).json({ message: 'page  not found' });
-});
-app.use('/*', express.static(path.join(__dirname,'..','..','public','index.html')));
-app.listen(process.env.PORT || 3000, () => {
-  console.log('server runs on 3000');
+const server = http.createServer(app);
+server.listen(process.env.PORt || 3000 , () => {
+  console.log('RUNNING ON 3000');
 });
